requirements.txt
numpy
Pillow
scipy
#komm